feat: add truck location

+++ application/src/main/java/fr/lespimpons/application/api/internal/controller/dto/TruckSensorDtoFromApi.java
@@ -2,6 +2,8 @@ package fr.lespimpons.application.api.internal.controller.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record TruckSensorDtoFromApi(@NotNull Long id, @NotNull double latitude, @NotNull double longitude) {
+import java.math.BigDecimal;
+
+public record TruckSensorDtoFromApi(@NotNull Long id, @NotNull BigDecimal latitude, @NotNull BigDecimal longitude) {
 
 }
